Improved tile display in Qt 5 GUI
The PinchMap element is was originally based on code from
the Advanced Geocaching Tool for Linux (AGTL) which (miss)used
the QTQuick image case in an ingenious way to display the tiled map.
The viewport was covered by a grid of tiles that was moved when the map
was panned.

If the pan did not uncover any not yet visible tile row or
column the grid was just nudged a bit. If a new row/column became
visible the visible rows/columns just switched image source url (with
one row/column address "falling of" and being replaced by the newly visible
row/column) and nudged the grid back a bit. Like this a constant number of tiles could
effectively simulate a slippy map.

Unfortunately this has drawbacks, some evident and some even
more horrendous that have only been discovered later:
- the QtQuick image cache sometimes gets flushed (application don't
  really have any direct control of it) which manifests as the map
  flashing due to visible tiles reloading from the tile image provider
- according to a discussion with woot on IRC (Thanks! :) ) the
  clever trick with image source url switching not only causes the above
  mentioned flickering but is also horrendously inefficient  as every
  image source url switch, even if the image is cached, requires
  new texture upload to the GPU (!!!!!), which is a very bad idea,
  especially for maps that can have tens of image, possibly in multiple
  layers

So to fix these issues, the "clever" url swapping trick has been
scrapped. The new solution works like this:
- there is a ListModel that track all tiles that should be visible at
  the moment
- a central repeater creates tile delegates based on this model
- each tile delegate is also a repeater with data model based on enabled
  map layers
- unlike the previous url swapping technique tiles are now dynamically
  added and removed from the visible tiles model as the map is panned,
  zoomed, recentered, etc.
- also as we now no longer need to care about image cache coherency
  we can also scrap the old horrendous "tile size notification" system
  that used tile size as a notification of tile availability and
  replace it with a more sane system based on callbacks and direct
  tile-downloaded notifications based on a tile-tracking map

The new way should have these benefits:
- better performance as tile textures don't need to be reuploaded to CPU
  all the time
- no loaded tile flickering
- more sane asynchronous tile loading